refactor: Remove actor tools module and integrate project tools for input mapping functionality
